Move `subscribeAndCount` test only utility out of main package
