+++ config/jest.config.js
@@ -12,7 +12,8 @@ module.exports = {
   testURL: 'http://localhost',
   testPathIgnorePatterns: [
     '/node_modules/',
-    '/dist/'
+    '/dist/',
+    'src/__tests__/utils'
   ],
   modulePathIgnorePatterns: ['/dist/'],
   setupFiles: ['<rootDir>/src/config/jest/setup.ts'],
