+++ src/__tests__/utils/subscribeAndCount.ts
@@ -1,7 +1,7 @@
-import { ObservableQuery } from '../core/ObservableQuery';
-import { ApolloQueryResult } from '../core/types';
-import { Subscription } from '../util/Observable';
-import { asyncMap } from './observables';
+import { ObservableQuery } from '../../core/ObservableQuery';
+import { ApolloQueryResult } from '../../core/types';
+import { Subscription } from '../../util/Observable';
+import { asyncMap } from '../../util/observables';
 
 export default function subscribeAndCount(
   done: jest.DoneCallback,
